+++ babel.config.js
@@ -4,7 +4,6 @@ module.exports = api => {
   return {
     presets: [
       "@babel/preset-env",
-      "@babel/preset-stage-3",
       "@babel/preset-react",
       "@babel/preset-typescript"
     ]