+++ stories.tsx
@@ -1,14 +1,13 @@
 import * as React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import BreakpointObserver from './index.tsx';
+import BreakpointObserver from './';
 
 storiesOf('breakpoint-observer', module).add('Current Breakpoint', () => (
   <React.Fragment>
     <h2>Observe Current Breakpoint</h2>
     <p>
-      Supplying a breakpoint object to <code>{`<BreakpointObserver />`}</code>{' '}
-      will return the current breakpoint.
+      Supplying a breakpoint object to <code>{`<BreakpointObserver />`}</code> will return the current breakpoint.
     </p>
     <pre>
       <code>
@@ -26,9 +25,7 @@ storiesOf('breakpoint-observer', module).add('Current Breakpoint', () => (
 storiesOf('breakpoint-observer', module).add('Default Breakpoint', () => (
   <React.Fragment>
     <h2>Render Server-Side With a Default Breakpoint</h2>
-    <p>
-      In environments without a window, a default breakpoint can be supplied.
-    </p>
+    <p>In environments without a window, a default breakpoint can be supplied.</p>
     <pre>
       <code>
         {`<BreakpointObserver breakpoints={{ mobile: 0, tablet: 768 }} defaultBreakpoint={'tablet'}>
@@ -36,28 +33,22 @@ storiesOf('breakpoint-observer', module).add('Default Breakpoint', () => (
 </BreakpointObserver>`}
       </code>
     </pre>
-    <BreakpointObserver
-      breakpoints={{ mobile: 0, tablet: 768 }}
-      defaultBreakpoint={'tablet'}
-    >
+    <BreakpointObserver breakpoints={{ mobile: 0, tablet: 768 }} defaultBreakpoint={'tablet'}>
       {breakpoint => <p>Current breakpoint is {breakpoint}!</p>}
     </BreakpointObserver>
   </React.Fragment>
 ));
 
 storiesOf('breakpoint-observer', module).add('Callback Function', () => {
-  const myCallback = breakpoint =>
-    console.log(`The current breakpoint is ${breakpoint}!`);
+  const myCallback = breakpoint => console.log(`The current breakpoint is ${breakpoint}!`);
 
   return (
     <React.Fragment>
       <h2>Callback function</h2>
       <p>
-        The <code>{`<BreakpointObserver />`}</code> can also be used with a
-        callback function that receives the current breakpoint. This can then be
-        stored in your application state. This way you do not actually supply
-        any children to <code>{`<BreakpointObserver />`}</code> and can just
-        include it once in your application's top level.
+        The <code>{`<BreakpointObserver />`}</code> can also be used with a callback function that receives the current
+        breakpoint. This can then be stored in your application state. This way you do not actually supply any children
+        to <code>{`<BreakpointObserver />`}</code> and can just include it once in your application's top level.
       </p>
       <pre>
         <code>
@@ -71,11 +62,7 @@ storiesOf('breakpoint-observer', module).add('Callback Function', () => {
 />`}
         </code>
       </pre>
-      <BreakpointObserver
-        breakpoints={{ mobile: 0, tablet: 768 }}
-        defaultBreakpoint={'tablet'}
-        callback={myCallback}
-      />
+      <BreakpointObserver breakpoints={{ mobile: 0, tablet: 768 }} defaultBreakpoint={'tablet'} callback={myCallback} />
     </React.Fragment>
   );
 });
@@ -84,9 +71,8 @@ storiesOf('breakpoint-observer', module).add('Max and Min Width', () => (
   <React.Fragment>
     <h2>Observe Current Breakpoint's Max and Min Width</h2>
     <p>
-      <code>{`<BreakpointObserver />`}</code> also returns the current
-      breakpoint's min and max widths as the second and third arguments,
-      respectively.
+      <code>{`<BreakpointObserver />`}</code> also returns the current breakpoint's min and max widths as the second and
+      third arguments, respectively.
     </p>
     <pre>
       <code>
@@ -98,9 +84,9 @@ storiesOf('breakpoint-observer', module).add('Max and Min Width', () => (
     <BreakpointObserver breakpoints={{ mobile: 0, tablet: 768 }}>
       {(breakpoint, minWidth, maxWidth) => (
         <p>
-          The current breakpoint is {breakpoint} with minWidth of {minWidth}px{maxWidth
-            ? ` and a maxWidth of ${maxWidth}px`
-            : ''}!
+          The current breakpoint is {breakpoint} with minWidth of {minWidth}
+          px
+          {maxWidth ? ` and a maxWidth of ${maxWidth}px` : ''}!
         </p>
       )}
     </BreakpointObserver>